Fixes to build on win32

+++ sscanApp/src/writeXDR.c
@@ -126,9 +126,9 @@ int writeXDR_opaque(FILE *fd, char *cp, int cnt) {
 	return (writeXDR_bytes(fd, zero, 4 - nPad));
 }
 
-int writeXDR_bytes(FILE *fd, char *addr, u_int len) {
+int writeXDR_bytes(FILE *fd, char *addr, size_t len) {
 
-	if ((len != 0) && (fwrite(addr, (int)len, 1, fd) != 1))
+	if ((len != 0) && (fwrite(addr, len, 1, fd) != 1))
 		return (0);
 	return (1);
 }
